requirements.txt
# ----------------------------
# Núcleo
# ----------------------------
numpy==1.26.4
pandas==2.1.4

# ----------------------------
# Machine Learning / Embeddings
# ----------------------------
torch==2.1.2
transformers==4.35.2
sentence-transformers==2.2.2
sentencepiece
huggingface_hub==0.16.4

# ----------------------------
# LLM (OpenAI/Ollama)
# ----------------------------
openai>=1.0.0

# ----------------------------
# Scraping
# ----------------------------
beautifulsoup4
lxml
requests

# ----------------------------
# MongoDB
# ----------------------------
pymongo

# ----------------------------
# Utilidades
# ----------------------------
python-dotenv
tqdm

# ----------------------------
# Web API
# ----------------------------
fastapi
uvicorn[standard]

# ----------------------------
# Scheduler
# ----------------------------
apscheduler

# ----------------------------
# Dependências auxiliares
# ----------------------------
annotated-types
anyio
click
colorama
exceptiongroup
h11
idna
pydantic_core
sniffio
starlette
typing-inspection
typing_extensions
